requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==3.2.2
djangorestframework==3.12.4
exceptiongroup==1.1.1
gunicorn==20.0.4
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-decouple==3.4
pytz==2023.3
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
